Changed the position parameters (x and y) to the double type

+++ src/main/java/spacewars/controller/game/PlayerController.java
@@ -29,8 +29,8 @@ public class PlayerController {
 //    }
 
     protected Position applyCollisions(Position position) {
-        int x = player.getPosition().getX();
-        int y = player.getPosition().getY();
+        double x = player.getPosition().getX();
+        double y = player.getPosition().getY();
 
         int s = player.getSpeed();
